feat(Action): Implement an action listener and stopper to prevent player from doing given task while doing other given task.

+++ pps/src/main/java/pewpew/smash/engine/entities/MovableEntity.java
@@ -15,7 +15,7 @@ public abstract class MovableEntity extends UpdatableEntity {
 
     @Setter
     private Direction direction = Direction.UP;
-    private int speed;
+    private float speed;
 
     @Getter
     private int prevX, prevY;
@@ -27,7 +27,7 @@ public abstract class MovableEntity extends UpdatableEntity {
         y += direction.getVelocityY((int) (speed));
     }
 
-    public void setSpeed(int speed) {
+    public void setSpeed(float speed) {
         if (speed < 0) {
             throw new IllegalArgumentException("Speed can't be negative");
         }
